Move: `WrongNetworkDialog` to proper location
