+++ src/modules/dashboard/components/Dialogs/WrongNetworkDialog/index.ts
